+++ src/main/kotlin/domain/player/BlackPlayer.kt
@@ -1,8 +1,8 @@
-package player
+package domain.player
 
-import Stone
-import state.PlayerState
-import state.PlayingState
+import domain.state.PlayerState
+import domain.state.PlayingState
+import domain.stone.Stone
 
 class BlackPlayer(state: PlayerState = PlayingState()) : Player(state) {
     override fun putStone(stone: Stone): Player = BlackPlayer(state.add(stone))
