+++ src/main/kotlin/domain/player/WhitePlayer.kt
@@ -1,8 +1,8 @@
-package player
+package domain.player
 
-import Stone
-import state.PlayerState
-import state.PlayingState
+import domain.state.PlayerState
+import domain.state.PlayingState
+import domain.stone.Stone
 
 class WhitePlayer(state: PlayerState = PlayingState()) : Player(state) {
     override fun putStone(stone: Stone): Player = WhitePlayer(state.add(stone))
