+++ src/main/kotlin/domain/state/PlayingState.kt
@@ -1,7 +1,7 @@
-package state
+package domain.state
 
-import Stone
-import Stones
+import domain.stone.Stone
+import domain.stone.Stones
 
 class PlayingState(stones: Stones = Stones()) : PlayerState(stones) {
     override fun add(newStone: Stone): PlayerState {
