+++ src/main/kotlin/domain/stone/StoneColor.kt
@@ -1,3 +1,5 @@
+package domain.stone
+
 enum class StoneColor {
     BLACK, WHITE;
