+++ src/main/kotlin/listener/OmokTurnEventListener.kt
@@ -1,7 +1,7 @@
 package listener
 
-import domain.Position
-import domain.StoneColor
+import domain.position.Position
+import domain.stone.StoneColor
 
 interface OmokTurnEventListener {
     fun onTakeTurn(stoneColor: StoneColor): Position
