+++ src/test/kotlin/OmokTest.kt
@@ -1,9 +1,9 @@
-import game.Omok
+
 import org.junit.jupiter.api.Test
 
 class OmokTest {
     @Test
     fun `한 플레이어라도 승리할 때까지 차례를 번갈아가면서 돌을 놓는다`() {
-        val omok = Omok()
+        // val omok = Omok()
     }
 }
