+++ src/test/kotlin/PositionTest.kt
@@ -1,3 +1,4 @@
+import domain.position.Position
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
