+++ src/test/kotlin/StoneColorTest.kt
@@ -1,3 +1,4 @@
+import domain.stone.StoneColor
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
