+++ src/test/kotlin/StoneTest.kt
@@ -3,7 +3,7 @@
 class StoneTest {
     // @Test
     // fun `오목알은 자신의 위치를 알고 있다`() {
-    //     val stone = Stone("H", 10)
+    //     val stone = domain.stone.Stone("H", 10)
     //     assertThat(stone.x).isEqualTo("H")
     //     assertThat(stone.y).isEqualTo(10)
     // }
