+++ src/test/kotlin/StonesTest.kt
@@ -1,3 +1,4 @@
+import domain.stone.Stones
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
